+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * `Data.Interval.Borel.isSubsetOf`
 * `Data.Interval.Layers.difference`
 * `Data.Interval.Layers.clip`
+* `Data.Interval.Layers.toStepFunction`
 * `Data.Calendar.happeningAt`
 
 ## 0.2.0.1 - 2022-05-30